# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Spacenews Flight API",
        "version": "1.0.0"
    },
    "paths": {
        "/api/articles": {
            "get": {
                "tags": [
                    "Articles"
                ],
                "summary": "Articles index",
                "operationId": "11aa4cc6561a7b94c56d1201d62513ee",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": true,
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/articles/": {
            "post": {
                "tags": [
                    "Articles"
                ],
                "summary": "Articles store",
                "operationId": "a81f7b4efd34044248ecc9726f844ac3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "url",
                                    "imageUrl",
                                    "newsSite",
                                    "summary",
                                    "publishedAt",
                                    "updatedAt"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "An article :)"
                                    },
                                    "url": {
                                        "type": "string",
                                        "example": "https://www.spaceflightnews.com/newarticle"
                                    },
                                    "imageUrl": {
                                        "type": "string",
                                        "example": "https://www.spaceflightnews.com/newarticle.png"
                                    },
                                    "newsSite": {
                                        "type": "string",
                                        "example": "Space Flight News"
                                    },
                                    "summary": {
                                        "type": "string",
                                        "example": "New article :o"
                                    },
                                    "publishedAt": {
                                        "type": "string",
                                        "example": "2022-05-16T22:43:44.148Z"
                                    },
                                    "updatedAt": {
                                        "type": "string",
                                        "example": "2022-05-16T22:43:44.148Z"
                                    },
                                    "featured": {
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "launches": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "id": {
                                                    "type": "string",
                                                    "example": "l4unc8-1d"
                                                },
                                                "provider": {
                                                    "type": "string",
                                                    "example": "Launch Provider"
                                                }
                                            },
                                            "type": "object",
                                            "format": "query"
                                        }
                                    },
                                    "events": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "id": {
                                                    "type": "string",
                                                    "example": "3v3nt-1d"
                                                },
                                                "provider": {
                                                    "type": "string",
                                                    "example": "Event Provider"
                                                }
                                            },
                                            "type": "object",
                                            "format": "query"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/api/articles/{id}": {
            "get": {
                "tags": [
                    "Articles"
                ],
                "summary": "Articles show",
                "operationId": "b3e529eab04efb5d305b07e2e58d9869",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Articles"
                ],
                "summary": "Articles update",
                "operationId": "2717d5a8e3b1f16637aded325347f9ad",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "url",
                                    "imageUrl",
                                    "newsSite",
                                    "summary",
                                    "publishedAt",
                                    "updatedAt"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Edited article :)"
                                    },
                                    "url": {
                                        "type": "string",
                                        "example": "https://www.spaceflightnews.com/editedarticle"
                                    },
                                    "imageUrl": {
                                        "type": "string",
                                        "example": "https://www.spaceflightnews.com/editedarticle.png"
                                    },
                                    "newsSite": {
                                        "type": "string",
                                        "example": "Space Flight News"
                                    },
                                    "summary": {
                                        "type": "string",
                                        "example": "Edited article :o"
                                    },
                                    "publishedAt": {
                                        "type": "string",
                                        "example": "2022-05-16T22:43:44.148Z"
                                    },
                                    "updatedAt": {
                                        "type": "string",
                                        "example": "2022-05-16T22:43:44.148Z"
                                    },
                                    "launches": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "id": {
                                                    "type": "string",
                                                    "example": "l4unc8-1d"
                                                },
                                                "provider": {
                                                    "type": "string",
                                                    "example": "Launch Provider"
                                                }
                                            },
                                            "type": "object",
                                            "format": "query"
                                        }
                                    },
                                    "events": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "id": {
                                                    "type": "string",
                                                    "example": "3v3nt-1d"
                                                },
                                                "provider": {
                                                    "type": "string",
                                                    "example": "Event Provider"
                                                }
                                            },
                                            "type": "object",
                                            "format": "query"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Articles"
                ],
                "summary": "Articles destroy",
                "operationId": "97f34592611a1dd53a086865e29706c9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
...
    }
}